[Development] Implemented `Vec` newtype wrapper
Implemented `Vec` newtype wrapper over `Vector`. This abstraction provides more mathematical functions. List of those could be changed in the future (note: some operations that are implemented on `Vector` are not reimplemented for `Vec`, because they are not as common and it is possible to unwrap `Vec` and perform those operations if you need them).