[Development] Implemented `Layers.Activations` submodule
Implemented `Synapse.ML.Layers.Activations` submodule which provides activation functions - unary functions that are differentiable almost everywhere and so they can be used in backward loss propagation.
Currently, there are not many of different activation functions implemented, but it is enough to serve as a proof-of-concept. More functions will be added later.