[Development] Implemented `Synapse.ML.Training.Metrics` submodule
Implemented `Synapse.ML.Training.Metrics` submodule which provides collection of functions that are used to judge the performance of models.
Currently, only conversion between losses to metrics is implemented, more metrices will be provided later in the development.
Imposed constraint on `LossFn` and `MetricsFn` which will be important when `fit` function will be implemented. If API will not need those constraints, those will be lifted. There is also a subject of change of `MetricFn` type, because some metrics are stateful.